Move test files into proper packages
