+++ src/test/java/item/ItemTest.java
@@ -1,4 +1,4 @@
-package seedu.moneygowhere;
+package item;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
