--- src/test/java/seedu/moneygowhere/DukeTest.java
@@ -1,12 +1,0 @@
-package seedu.moneygowhere;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-class DukeTest {
-    @Test
-    public void sampleTest() {
-        assertTrue(true);
-    }
-}
